Initial cf setup (#50)
