+++ Gopkg.toml
@@ -50,12 +50,16 @@
 
 [[constraint]]
   name = "github.com/pmorie/go-open-service-broker-client"
-  version = "0.0.5"
+  version = "=0.0.5"
 
 [[constraint]]
   name = "github.com/pmorie/osb-broker-lib"
-  version = "0.0.6"
+  version = "=0.0.10"
 
 [[constraint]]
   name = "github.com/gavv/httpexpect"
   revision = "c44a6d7bb636b17e880a53998a7f7061a56ffacb"
+
+[[constraint]]
+  name = "github.com/cloudfoundry-community/go-cfenv"
+  revision = "e84b5c116637ef174eb8c7d68a28bc7269b3b9cb"
