+++ application.yml
@@ -1,9 +1,10 @@
 server:
   requestTimeout: 3000
   shutdownTimeout: 3000
-  port: 8081
+  port: 8080
 log:
   level: debug
   format: text
 db:
-  uri: postgres://postgres:postgres@localhost:5555/postgres?sslmode=disable
+  name: sm-postgres
+  uri: postgres://postgres:postgres@192.168.99.101:5555/postgres?sslmode=disable
