+++ cmd/cf/main.go
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2018 The Service Manager Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package main
+
+import (
+	"context"
+	"os"
+	"os/signal"
+
+	cfenv "github.com/Peripli/service-manager/cmd/cf/env"
+	"github.com/Peripli/service-manager/sm"
+	"github.com/sirupsen/logrus"
+	"github.com/Peripli/service-manager/env"
+)
+
+func main() {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	handleInterrupts(ctx, cancel)
+
+	srv, err := sm.NewServer(ctx, cfenv.New(env.Default()))
+	if err != nil {
+		logrus.Fatal("Error creating the server", err)
+	}
+	srv.Run(ctx)
+}
+
+func handleInterrupts(ctx context.Context, cancelFunc context.CancelFunc) {
+	term := make(chan os.Signal)
+	signal.Notify(term, os.Interrupt)
+	go func() {
+		select {
+		case <-term:
+			logrus.Error("Received OS interrupt, exiting gracefully...")
+			cancelFunc()
+		case <-ctx.Done():
+			return
+		}
+	}()
+}
