+++ cmd/cf/manifest.yml
@@ -1,0 +1,8 @@
+---
+applications:
+  - name: service-manager-core-test
+    env:
+      GOPACKAGENAME: github.com/Peripli/service-manager
+      GOVERSION: go1.10
+    services:
+      - sm-postgres
